Label Position Optimizations
For the initial vector label layout, I matched the position of the labels pixel-perfect to the original layout I made over in Photopea. In this update, I'm adjusting/removing a few extra values to make things more consistent to work in the web world. Since the text is first class over in SVG, it is handled a bit differently in terms of it's box shape, and in Photopea it was treated more as a path when I was moving/centering/positioning it on the layout. That made it a little less accurate in regards to things like spacing away from the edge of the letters, and this helps fix that.