Vector Labels!
Mostly got it working! The only holdup is getting Noto Sans to load. For standalone SVG images, they aren't allowed to load external resources, which is something that Inline SVG *can* do (when it's directly a child of the page's HTML). I found a workaround online that will probably have the solution to make it work. You simply just have to embed the font as a Data URL inside of the SVG itself. It's kind of a bummer that it may make it slower, but that's pretty much the only option available. The other thing I haven't tried that may work is if you simply install the font to your system, then it could load it that way. So I think it's down to either embedding the font as Base64 into the SVG source, using a different font (a common system-supplied one or something), installing the font to your system, or I could move away from SVG for layout, and try to figure something else out. While that would probably suck to handle, it is another option. I like the fluidity that the CSS provides though, so I think I will probably end up taking that over the font situation.

https://alligatr.co.uk/blog/render-an-svg-using-external-fonts-to-a-canvas/