Full-Vector Thumbnail Demo
Figured out how to fully move the thumbnail setup into SVG! Alog with this, I also finally found a way to get a Gaussian Blur, and have solid edges! It uses a transform effect on the output image that maps all pixels to be fully opaque. I love how simple this is, and with testing, the blur and the transform effect both work in Safari too!!! So, now the next part is getting to render this on the canvas itself, adding the artwork and font into the static SVG content (dynamically, since each thumbnail is different) so it can be a self-sufficient standalone image (required to load an SVG in a canvas, will probably be with Data URIs). And once that is essentially "working", then I can start testing around with WASM ffmpeg! This project is getting more awesome...

Amazing help from StackOverflow:
https://stackoverflow.com/questions/24416564/svg-blur-filter-edge-artifacts