Data URL: Artwork Embed!
Got the artwork to embed directly into the SVG drawn to the canvas! I tried doing this the other day, and something was going wrong, and I'm not sure what it was, but it worked the first try this time around! Next I can use this same function to embed the font for the labels, the new function's meant to work with any old Blob, not only just images.

I was suprised to see that it worked to embed the artwork, on the first reload! I went back to my Live Server tab with the site up, and it was already refreshed with the expected result. Wow!

I love that power behind Promises and async-await now. It's so much better for me to conceptualize how the asynchrony works. I think it's a mix of that, and now knowing how to manage multiple concurrencies at once a bit better. Now it's much more familiar to me, and I can see how things should line up together. I like to think of each task like a timeline, where you just have to make sure line up the start and end points of each task, then they have the ability to access each other's results as needed.

Today's music playlist:
Some of my own (new) songs, listening back to work on any mixing that needs some work. (Road to Forget album, not released yet)
Mike Keneally - The Scambot Holiday Special
Marco Minnemann - The Ugliest of Beasts
King's X - Dogman (Album)
Porcupine Tree - The Sky Moves Sideways (With the extended songs, like Moonloop (listening to that now :) ))