Embed CSS URLs: For Font Faces!
Did some extra fancy asynchronous fetching and string manipulation to embed the Google Font font files as Base64 into the SVG's CSS font declarations stylesheet!

So, essentially I will eventually inject the Google Fonts URL into the SVG's `innerHTML` as a string, after running `embedCSSURLs()` on it's text content after being fetched. I could've just manually embeded the font file for just English (and some other languages) into the SVG's source, as a Base64 URL, but I wanted it to work with Google Fonts' dynamic nature, since I think they can update the font files automatically on their end, and I'd want to keep that availability. I also wanted to try my best to set up the ground work for being able to configure how the SVG looks down the line, and getting stuck in my own implementation details to add that wasn't something I'd like to run into, the best I can at least.

This works really well! The only thing that you'd need for this to work on a larger scale, outside of this project, is to handle things like other CSS files, like with using the `@import` statement, and rescursively into those as well. I'm at the very minimum glad that this is already producing a valid stylesheet composed of dynamic Data URLs for the linked content, and I wanted to make sure to snapshot this stage while it is working, before I test it on the actual font system for the SVGs. That's coming up here, right after this!