Full-Vector Thumbnail: Custom Fonts!
Got embedding the Noto Sans font into the static SVG working!!! Super happy with this already. The thumbnail setup is now essentially feature complete, so now it's time for some code reworking to organize things out, and remove any rough edges that are still around. I am debating moving this project to TypeScript too, but I'm not totally sure yet. I like having the ability to define interfaces in TS, and it looks like that's not quite as simple to make with JSDoc. JSDoc has been perfect so far for variable type inference and function declaration types, but it does feel like it gets a little more hectic when you start making an outwards-facing API, which is something I will want to have for the project also. I think I'll probably hold myself from moving to TypeScript until I really *need* to make an API. This isn't ready to provide others with something to build with yet, and that's mainly what I'd want to have TypeScript for eventually. So, yeah!

Ran a half marathon today, it was very nice out! A little chilly, but sunny, so the right combo :)

Forgot to add in my other recent commits (I think?), I've started looking into Frank Zappa, and MAN, was he a cool guy! Can't wait to listen to his albums, as I already love Keneally and Minnemann's projects, and those tend to be pretty far out there too. I *strongly* love music now.

If you are a fellow music nerd too, feel free to send any suggestions my way, and I'd be happy to do the same back to you!