Tainted Canvas
Tried calling the export methods for the `<canvas>`, and it threw a tainted image data error! Turns out that there is a limiter that prevents you from loading an SVG with a `<foreignObject>` tag in it. It is explained very well here:
https://stackoverflow.com/questions/40897039/problems-with-getting-canvas-datauri-from-svg-with-foreignobject

Using the knowledge from that article, I made a workaround that follows a similar path, and it creates a Data URL for the SVG that is plain text based, rather than Base64, as for some reason only the Base64 URL types restrict the use of the `<foreignObject>` tag, but the escaped string version doesn't. How interesting!

At first I thought that my embedded artwork images were flagging the canvas as tainted, but removing those didn't fix it. It makes sense how the HTML document shouldn't be able to be displayed, in a way. I wish it was a more available feature though, as it works really nicely for dynamic styling and positioning, like with using Flexbox.

This workaround appears to work nicely in Chrome, most of the time in Safari, and not working on Firefox unfortunately. If I can use an alternative without the `<foreignObject>` setup, I think I will, as that seems to be the most complex part of this so far. It will make the output SVG more widely compatible for non-browser things too, and that could be a great benefit actually! Definitely wanted to commit this working version (like I've been doing with the other updates), as this was a fairly complex spot to get into having it working. So glad it's at this point, this is already really neat!

(I may use this one to replace my current `<foreignObject>` setup to make things a bit simpler, and more stablily supported.)
https://stackoverflow.com/questions/45521822/svg-center-dynamic-text-within-svg-elements-blocks