Data URL Embed Functions
Expanded `toDataURL()` into two separate functions! Now one is dedicated to simply plain text entries only, and the other is for Blobs to Base64. They are `toDataURLComponent()` and `toDataURLBase64()`.

I'm going to try and break down this crazy "spec" of getting the SVG/ForeignObject/Canvas madness down to distinct parts, then I can rewrite it knowing fully how it works, and what double-negatives I can remove (not sure if those are what's in there specifically, but I came across that a few updates ago. One example was that I was converting a `fetch()` into a Blob, only to turn it into a string again, or something along those lines).

I tried out making the layout without the ForeignObject tag and the embedded HTML, but it didn't seem to come to easy to get the text wrapping and centering that I want to have (have already! :O). So, I think I'm gonna stick with this, hopefully it doesn't break from a browser perspective, since it's a bit finnicky to render ForeignObjects to a Canvas, and still be able to read the content afterwards. It seems to work, but feels very loophole-ey.

After trying that out, I ended up getting these changes working, after reading up a bit more on how you can get it to work (ForeignObject rendering). This link was a little refresher to read, it seems a little out of date though? I already have it working here, so I didn't want to change what I had to see if theirs worked. A great find to hear that they got it working though! That gives me hope that it's not *quiiite* so finnicky (at least, hopefully it will stick around as time goes on, eeeeek), since someone else also managed to get the same successful result as I did. The age of the article makes me a bit skeptic of if things have changed since then though.
http://man.hubwiz.com/docset/JavaScript.docset/Contents/Resources/Documents/developer.mozilla.org/en-US/docs/Web/API/Canvas_API/Drawing_DOM_objects_into_a_canvas.html