FFmpeg Encoding Test!
Figured out how to encode my generated thumbnail alongside my song audio, into a video using FFmpeg! Since this is just in the test stage essentially, I'm just adding some links to how I got it to work, and a shell script of what I wrote in the terminal to run it. Now that I have a documented working version of this using `ffmpeg`, I will look into doing the same thing, but in the browser using FFmpeg WASM! It will be nice to have TS types for the API too, as I'm not completely 100% about what all the flags do. I think I understand most of them though:

ffmpeg - Starts the command (hehe)
loop - boolean, loop the next input with an infinite length, as needed
framerate - sets the framerate to 1 (this happens in multiple spots?)
i - first source, this is the image, 26.png
map 0 - get the first (only in this case) source value entry from the first input
map 1:a - get the audio only, from the second input source
c:v libx264 - I think this defines the codec for the output video format
preset: ultrafast - Not completely sure, I think this makes it run faster? I'm curious if this degrades the quality though, so I'm gonna experiment with adding/removing this to see what changes
tune: stillimage - Didn't look into this yet, but it seems like it probably makes the encoding more efficient if you know that the video is just one image the entire time
vvf - Video effects, sets the FPS to 10 (second time around?), sets the video format to yuv420p, not quite sure what that is yet
c:a copy - ooop! Hopefull this means make a simple copy of the audio, without reencoding it? That would be perfectly what I'm looking for! I want to keep the quality as high as possible, since this is for music specifically.
shortest - This sets the length of the output video to be the length of the shortest input source. In this case, the image can expand infinitely, so it's essentially "inifinite", while the audio for 26.m4a is 7 minutes. Since that is the shorter one, the video will be 7 minutes long, with the image truncating to fill that space specifically.
26.mp4 - The output file name!

This is nicely simple, and complex. Definitely feels very terminal-worthy. Like working with Bash, after getting used to hearing all of these different flags and abbreviations, it's gonna be super powerful to specifically critique your video format exactly to how you want it to be exported! I love how extensive `ffmpeg` is, it's mind-blowing how much people have put into it. This is inspirational to how I want my Minecraft world converter/editor project/thing to eventually work; To just let the program do the heavy lifting, and you can work with the data directly, using it's APIs :)

Love this!