FFmpeg WASM (Experimental)
Additions:
- Added experimental support for FFmpeg WASM! This is a very hacky workaround to get set up, at the moment.

A few notes about how it works so far:
- FFmpeg WASM requires the use of `SharedArrayBuffer`, so I had to ensure that `window.crossOriginIsolated` resorts to true. To do that, your site has to be in a secure context, and all of it's assets have to be served with the `Cross-Origin-Opener-Policy` header set to `same-origin` (protects your origin from attackers), and the `Cross-Origin-Embedder-Policy` set to `require-corp` (protects victims from your origin). This is because of the 2018 Spectre vulnerability, and it is a new secure approach to start re-enabling shared memory, with security first in mind.
https://stackoverflow.com/questions/68616601/how-to-define-sharedarraybuffer-in-chrome
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/SharedArrayBuffer

Because this is a server header required change, this means you have to have access to change the headers of your resources sent to your page. This isn't available to modify for GitHub Pages, and it's still a bit out of my technical reach know how (at the moment :) ), so I looked more into it, and turns out that you can get around this by adding a Service Worker to your page, and it will allow you to change the headers for your fetched resources, and add the additional headers where needed. I got to try setting up a TS-safe Service Worker again with this too (bonus!), and it works *fairly* nice! Not quite just right yet, in terms of type-safety and a dev setup, but it does work as intended for what it's supposed to do! These pages really helped out with adding the headers to the files, and it did make it so I can get `SharedArrayBuffer` using `live-server` on `localhost`! I haven't debugged it yet, but the `live-server`-hosted site wasn't able to run, for some reason, on Safari (iOS/iPadOS). So, not sure where the error is for that yet. It works locally on the same machine, in Chrome though!
https://stackoverflow.com/questions/72603171/sharedarraybuffer-github-pages-and-dns
https://github.com/gzuidhof/coi-serviceworker/blob/master/coi-serviceworker.js
https://github.com/ffmpegwasm/ffmpeg.wasm/blob/master/examples/browser/transcode.worker.js

FFmpeg runs relatively quick in the terminal on my Mac, using it as a native counterpart. In the browser on my Mac, using WebAssembly, it was a lot slower unfortunately. Not sure what the problem is for that, maybe the threading or something. And, as I mentioned, something about it errored out on mobile Safari, so it doesn't appear to be the most consistent in the browser. After I'm gonna do some more testing, I'll decide whether running it in the browser is worth it enough, as it doesn't seem to be consistent to implement, or on par in terms of performance. So, the video generation part may be a Node-only script if things don't go quite the right way. I love the speed and dynamic nature the thumbnail generation part, in the browser, so I think that will almost definitely stay there. It kind of sucks to de-couple the project into both browser and Node, rather than just the browser, but I want to take performance over having it in one place. Maybe a bundled Desktop-only Electron app, could be the way to go after all? I kind of don't like the heaviness of that, but that would be easy to distribute and easy to setup from a user-perspective. A web-first PWA would be the *better* option though, that's even better :) So, hopes to getting it to work on-par to native-in the browser :D

I also feel like mobile Apple devices are more than capable at running that, so I want to try my best to not cut them out of the picture, as it would be great to just use your phone (or iPad) to generate the video, since they are definitely powerful enough to encode and edit videos in iMovie, soo... Almost there!! At least it *works* in Chrome. I thought I hit a roadblock at the `SharedArrayBuffer` part, but that workaround successfully made it work :)

**Mammoth WVH**