FFmpeg Terminal
Did some testing with FFmpeg WASM in the browser, and it's not quite up to par with what I would like in terms of performance. Using regular old FFmpeg in the terminal by itself works very snappy for this project, but using it in the browser on my same Mac was noticeable slower. Not sure if it's a configuration issue that I haven't figured out yet, or if it's just plain that much slower, but I think I'm gonna try making the video production part with Node JS instead. Or, maybe even just a console app. Probably with Node though :)

I also had to do a lot of hacks to get it work right in the browser, because of some requirements to be able to use `SharedArrayBuffer`. Even with those wonky fixes, which involved the Service Worker, it still didn't work in Safari for some reason. I'm not sure if the performance would be very great on mobile with WASM, if it was slower on the Mac. I would like to be able to do it from anywhere, but I don't want it to be a lot slower just because of that.