User Upload Inputs
Starting to work on the actual GUI for the site now! Now you can upload multiple song files, and it will run the generator on them, and make thumbnails using the canvas. Currently, it only logs the resuting image `Blob`s to the console. I'm gonna make a save/view screen that appears after they are generated, next.

I did all of these changes and additions yesterday afternoon, but didn't get a chance to commit them before having to leave for work.