Download Issue
Discovered that you can't right-click and download an image loaded from a Blob URL, if the Object URL you made has already been revoked. Seems fairly straightforward, but I was curious if it could still work, even with revoking the URL. The image is still displayed on the page after revoking the URL, so that was my semi-guesswork there. The more you know!

Another thing is that if you do this using the anchor tag Blob URL trick, clicking the link and immediately revoking the URL doesn't cause an issue, so I was suspecting that it *might* do the same. I think that probably works because it's might synchronously be loading the file downloader system prompt before the URL is destroyed. Not sure if that's the case, but it explains why that works, unsupsectingly. It also means that maybe I shouldn't rely on that working! Sounds a bit insecure to me, if say it were to allow the Blob URL to be destroyed before showing the prompt.

For the time being I'm not going to revoke the URLs, since they currently have no way from being removed from the UI anyways, as the only way to reset the UI is to refresh the page. That also resets the Blob URLs, haha.