Full TypeScript!
This site is going to be in TypeScipt, just like all of my other projects haha.

Planning on some new features coming up here, and this will help out a lot with getting those working faster. The setup for Dovetail has made my site development a lot easier to build now! Being able to have GitHub Pages and TypeScript is a huge win.