TS Syntax
This simply moves over the current JSDoc type definitions for the app to TypeScript's syntax. This makes the app fully TypeScript now! JSDoc types are a solid recommend from me, this is how fast it was to move to TypeScript from the initial JavaScript codebase. You just have to have type-safe code before you are in TypeScript, then it's a breeze to move to it. So, the rule of thumb in the end, add your types as soon as possible!