Improved GUI Error Handling
Heard in #1 that the only way that PiSaucer found out what happened when the error occurred when opening a WAV file was that they had to open DevTools to find the error.

Now it's easier to see what errors are occurring in the app when you are in the browser, and you may not have access to DevTools at that given time.

This also restores the GUI so you can try more song files again, without having to refresh the page, just to reset the loading GUI back to a useable state.