Vite Tooling
Bringing this setup over from my other projects.

I think I may start fresh with this project, and maybe make it a terminal-only thing, I'm not totally sure. This one does work well, but sadly it only seems to work effectively in Chromium browsers. Not sure if it's based on a bug or something (I had to get some hacky things set up to get the HTML rendering working on the canvas), but I think making it built with Puppeteer in the end is probably the more streamlined way to go. Moving it over to being a Node-only app can also help with the addition of the YouTube uploading part, which is half of the design plan for the app. It has sounded not very UX-friendly to have both browser and Node counterparts to sections of what the app does, so maybe just moving it over to Node-only can improve it's usability. I would have liked for it to work on all platforms, but considering neither the video or image generation processes are nicely working in the browser, I can for now at least decide to go the other way, and try with what has a more likely chance of working. Moving to Node can also embed the video generation step too, as that's already a terminal-first thing, and it's working great! I'm curious if my progress on jsmediatags has been helpful or not, so far too, I think I'm going to work with that a little more again, or maybe try to use only ffmpeg to read the song metadata, as I think it also has the ability to do that. If it's going to be a terminal-first app instead, it could make sense to just use ffmpeg since that will be part of the program either way, for the video production step.

Tried adding jsmediatags as a regular npm dependency that Vite can add to the bundle, but jsmediatags is still not capable of being loaded from a plain ESM context for some reason. Not that is has exports, but it actually errors out, I think because module force strict mode, which jsmediatags doesn't appear to be able to handle. My TypeScript fork will likely solve any of those issues, maybe it's just the npm build? :)