Exp-Crazy-Rework
Doing all of the things, moving this project over to a CLI only setup. Going to get YouTube uploads, video generation, image generation, and song parsing all working under a simple to use CLI API, similar to that of ffmpeg. This is a big project that powers a small niche, sounds perfect for me lol.

https://pptr.dev/#usage
https://developer.chrome.com/docs/puppeteer/get-started/
https://quanticdev.com/articles/automating-my-youtube-uploads-using-nodejs/ (I've had this saved to my favorites for months now)
https://www.freecodecamp.org/news/how-to-use-puppeteer-with-nodejs/
https://stackoverflow.com/questions/20643470/execute-a-command-line-binary-with-node-js (yay!)
https://youtu.be/y_-iFkfGLWA?t=325
https://x-team.com/blog/a-guide-to-creating-a-nodejs-command/ (extra yay! `npm link` is an absolutely killer find, it works with `tsc -w` too :O )
https://stackoverflow.com/questions/20638520/appropriate-hashbang-for-node-js-scripts
https://www.digitalocean.com/community/tutorials/nodejs-command-line-arguments-node-scripts
https://www.npmjs.com/package/commander