Web-App-Next-Step
Worked on these changes yesterday, didn't get a chance to commit them yet. I had been experimenting with making this a CLI-first tool again, and it wasn't working out as much as I'd thought. I still want to have that eventually, but I think settting things up to work GUI/web-first will make it easier to use it in a Puppeteer/Electron environment, rather than doing everything through the terminal specifically. The only thing I think I would have to do for this current code to make that version doable, would be to render the thumbnail using plain HTML, rather than SVG on the canvas (it would just be simpler/more reliable), then I could use Puppeteer to capture the rendering of that on the window.

I'd like to debug getting this to work in other browsers, when using the current SVG rendering method, but I think it might be a dead end thing (maybe, not giving up though), as I think Chromium is the only place it's working as of now. So I think it might be relying on a bug that *shouldn't* allow it to work?

Moving this to CLI-only, and using Electron/Puppeteer rendering using plain HTML (Chromium isn't a requirement when moving to the HTML rendering, can you use Tauri with Puppeteer??) would be the nicest way to build the CLI on top of I think, I'd like to have it working using that setup.

I tried looking into Skia-Canvas for this, but it didn't have support for rendering SVGs unfortunately, so that's out, as I need to be able to render the text using flexbox (I can do that handling programatically, but having that delegated to plain CSS logic seems more straightforward to me; less things to break unexpectedly), which can only be done inside of a `<foreignobject>` element.

This update also brings some other niceties that I've accumulated from working on my other projects.

I'm looking more into something like Preact/React for my projects, just learning a bit more about how to set it up properly. I think it could be a great use-case for the SVG/HTML tags for the Thumbnail element/component! Right now it's just made with an `innerHTML` string, but I think using JSX for that could be great, since you'd only need a single component too, you don't have to make a new one every time to render it.

Google Fonts npm?