Hacky Puppeteer!
Wow, this is sick! Got this set up using Puppeteer, and an HTTP server to host the site on localhost, which will be where the HTML is rendered to. Is there a better way of doing this that I haven't thought about yet? This is working out, but it does seem a bit ugly to have it working like this 😂

`tsx ./test`

This commit name is absolutely going to be a song name of mine.

I think this could indeed be a neat Electron app too, if all of this was done through a GUI, you could probably use the Electron backing for Puppeteer too right? I've seen a few sources and example of being able to do that. So you could have your app GUI run in Electron, and it would use that same implementation for the renderer window too. I like the CLI approach personally though, so I'm going to experiment with that to start with.

I think the GUI-app way is probably easier for more general users though, so I think that could be more worthwhile. It might make the renderer/Node interations a bit simpler too, as you'd only have to have the GUI bit, and you could use Context Bridges from Electron to get the Node communication set up? That worked fairly nicely in my now classic Electron Text Editor project.

Yeah, Tauri could be nice for this, but that's a bonus thing to worry about later. Having something working to start with is the initial goal yet again for this project here :)

https://stackoverflow.com/questions/47587352/opening-local-html-file-using-puppeteer
https://flaviocopes.com/node-serve-html-page/
https://stackoverflow.com/questions/46745014/alternative-for-dirname-in-node-js-when-using-es6-modules
https://www.bannerbear.com/blog/how-to-take-screenshots-with-puppeteer/