Demo: New Idea: SSR!
This train ran out of tracks, so I'm going to try taking a different line.

Realized that if I'm going to use an HTTP server, and I plan on using JSX, what if I just did this like a regular person and did this on the 'server' side? Lol

I was trying to load the file using Node's FS, then pass it into Puppeteer's page using `page.evaluate()` but turns out you can't serialize TypedArrays, Files, Blobs, or ArrayBuffers. The only way around this with that current setup would be to convert it to a Base64 DataURL, and that just started to make me feel yucky inside. There's already so many ugly steps going on here, and adding another layer of abstraction like that to simply get the data into Puppeteer (and it would have to come back out this way too) seemed not worth it to me. It started to wreak like the need for a rethink/write.

Took an hour or two to regroup (these changes were from before that), have a late dinner, play some music, and now I feel a bit better about things again. Lets's see where this goes next! Hell yeah!

THE NAME: Is the Primeagen

https://stackoverflow.com/questions/66711330/issue-listening-for-custom-event-via-puppeteer
https://stackoverflow.com/questions/63913846/how-do-i-call-function-outside-async-function-in-puppeteer-with-nodejs
https://stackoverflow.com/questions/46088351/how-can-i-pass-variable-into-an-evaluate-function
https://github.com/puppeteer/puppeteer/issues/3722
https://github.com/puppeteer/puppeteer/issues/2427
https://stackoverflow.com/questions/6850276/how-to-convert-dataurl-to-file-object-in-javascript
https://www.youtube.com/watch?v=ML743nrkMHw (Unrelated, but I watched part of it during my dinner break. Electron really is neat, I think I should give myself some slack (pun unintended), it really is powerful how many neat things you can do with it. Just because it is bulky compared to other options, it is really cool what it does provide you with things to do with it :) )