[Hacky] But Very Close!
Wow! Got this working 'fully on the terminal', thanks to Puppeteer! Of course, it's using Puppeteer under the hood, but now you can actually pass in your own song files, and it will programmatically return back to your what the resulting image is! So you don't have to manually work with the GUI like before.

Been listening to Thank You Scientist 'Terraformer', this album is so sick.

Now that this is working, I want to improve the API so it more simply does the handling, and you do things more functionally. Right now it makes a new Puppeteer browser window for every thumbnail generator, but it doesn't have to be that way, I can set it up much nicer.

I want to make it like the builder pattern for NBT reading and writing, where there are top-level functions that are for the user's ease of use, and to manage the handling on the backside of things so it only has one browser instance at a time. This will work super great with async-await. I also want to only have a single HTTP server at a time too, so I have to move where the definitions for each of these things are into the functions differently, allowing them to manage multiple thumbnail generations concurrently from one instance.

Thought I was going to need to get the MIME type for the artwork file, but I forgot that it's stored in the data itself, so JSMediaTags will provide it (I wrapped those into a Blob object). But if I didn't have that info available to me, I found a really great ESM library to handle that kind of logic.

https://flaviocopes.com/node-serve-html-page/ (Came back to look at this again, didn't need Vite/Express, rebuilt the renderer to simple use HTML templating instead!!! No need for the Canvas implementation now either, since you can just screenshot the window's content. So now you can demo out how the HTML/CSS looks for the thumbnail content, by just opening it with either Live Server, or STE! Yay :D)
https://stackoverflow.com/questions/4720343/loading-basic-html-in-node-js (Help with writing the HTML to the response object)
https://dev.to/victrexx2002/how-to-get-the-mime-type-of-a-file-in-nodejs-p6c
https://github.com/sindresorhus/file-type
https://developer.mozilla.org/en-US/docs/Web/HTTP/Basics_of_HTTP/Data_URLs
https://stackoverflow.com/questions/54099802/blob-to-base64-in-nodejs-without-filereader

I also brought up the updated JSMediaTags module I made, from the CLI branch. It had some better handling compared to that of the original 'site' implementation.