Modules! Timed-out, though
Got the modules set up! Now you can use the library using a single main function to start a generator/renderer/server. It looks like it's having some issues with loading times though, so it's preventing the screenshot from being taken, because it waits for the loading to finish, while it never does.

I think I might just go back to a normal class definition after this, Not sure if the modules make complete sense for this.

https://stackoverflow.com/questions/70009299/http-server-respnd-with-an-output-from-an-async-function (Looks like it might have to do with async-await, or maybe the tags global, or error handling? Not totally sure)