Arg Demoo
Between this and the last commit, found a reference as to just adding the `chmod +x` call to the build step itself. That is a possible idea, but I don't like that it won't work on Windows/non-Unix systems. I like that my projects can not only be used on any platform, but I want them to also be develop-able on any platform! I think the two should go hand in hand, at least in the nature of when it is something that should work cross-platform.

https://stackoverflow.com/questions/44612118/typescripts-transpiles-files-which-are-non-executable

Set up a simple CLI argument set using Vercel/Arg! Doing it on my own isn't going to be feasable, it is hard to catch a lot of the possible errors. For cases like these, I definitely do agree that using libraries are the way to go.

```sh
art-gen --input ./test/26.m4a --output ./demo.png
```

Right now you can also pass in multiple strings after the `--input` flag, and a match number of strings after the `--output` flag, providing for multiple files to create thumbnails for. I plan on adding additional arguments for things like specifying raw media tag values manually, which could allow you to generate the files using your own data, rather than reading it from the song directly. This will make working with WAV files possible, as you can provide what metadata you want to use for the thumbnails of those, and such.